pyproject.toml
[tool.poetry]
name = "lz-manager-service"
version = "0.1.0"
description = ""
authors = ["HezinTUKE <vova.gezin@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = ">=0.34.2,<0.35.0"
pyyaml = "^6.0.2"
sqlalchemy = "^2.0.41"
isort = "^6.0.1"
black = "^25.1.0"
alembic = "^1.16.1"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
elasticsearch = "^9.0.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"